Fix to apply child care eligibility to regular transactions
Prevent regular transactions of the child care category
from incrementing/decrement totals and all_sources inline
with similar behaviour for bank and cash transactions.

This was an ommission of required logic identified by
EFE team.

Child care costs (outgoings) should only ever be applied
if the applicant has dependants under 16 AND is employed
or is in receipt of a student loan.